[Feature] [~] = updating the header so that the packetnumber takes 2 bytes, modified all the methods in the decoder and encoder to accomodate the changes

+++ src/protocolHandler/Decoder.hh
@@ -17,7 +17,7 @@
 * the first parameter is the type of the packet and
 * the second is the number of the packet
 */
-#define HEADER std::pair<RType::PacketType, std::uint8_t>
+#define HEADER std::pair<RType::PacketType, std::uint16_t>
 #define UNKNOWN_HEADER std::make_pair(UNKNOWN, 0)
 #define PACKET std::pair<HEADER, COMMAND_ARGS>
